minor: move helper for setting up AAS settings in tests
